Move server package to dhcp6server
