+++ cmd/dhcp6d/main.go
@@ -1,4 +1,4 @@
-// Command dhcp6d is an example DHCPv6 server.  It can only assign a
+// Command dhcp6d is an example DHCPv6 dhcp6server.  It can only assign a
 // single IPv6 address, and is not a complete DHCPv6 server implementation
 // by any means.  It is meant to demonstrate usage of package dhcp6.
 package main
@@ -12,7 +12,7 @@ import (
 
 	"github.com/mdlayher/dhcp6"
 	"github.com/mdlayher/dhcp6/dhcp6opts"
-	"github.com/mdlayher/dhcp6/server"
+	"github.com/mdlayher/dhcp6/dhcp6server"
 )
 
 func main() {
@@ -34,7 +34,7 @@ func main() {
 
 	// Bind DHCPv6 server to interface and use specified handler
 	log.Printf("binding DHCPv6 server to interface %s...", *iface)
-	if err := server.ListenAndServe(*iface, h); err != nil {
+	if err := dhcp6server.ListenAndServe(*iface, h); err != nil {
 		log.Fatal(err)
 	}
 }
@@ -47,7 +47,7 @@ type Handler struct {
 
 // ServeDHCP is a dhcp6.Handler which invokes an internal handler that
 // allows errors to be returned and handled in one place.
-func (h *Handler) ServeDHCP(w server.ResponseSender, r *server.Request) {
+func (h *Handler) ServeDHCP(w dhcp6server.ResponseSender, r *dhcp6server.Request) {
 	if err := h.handler(h.ip, w, r); err != nil {
 		log.Println(err)
 	}
@@ -55,10 +55,10 @@ func (h *Handler) ServeDHCP(w server.ResponseSender, r *server.Request) {
 
 // A handler is a DHCPv6 handler function which can assign a single IPv6
 // address and also return an error.
-type handler func(ip net.IP, w server.ResponseSender, r *server.Request) error
+type handler func(ip net.IP, w dhcp6server.ResponseSender, r *dhcp6server.Request) error
 
 // handle is a handler which assigns IPv6 addresses using DHCPv6.
-func handle(ip net.IP, w server.ResponseSender, r *server.Request) error {
+func handle(ip net.IP, w dhcp6server.ResponseSender, r *dhcp6server.Request) error {
 	// Accept only Solicit, Request, or Confirm, since this server
 	// does not handle Information Request or other message types
 	valid := map[dhcp6.MessageType]struct{}{
@@ -166,7 +166,7 @@ func handle(ip net.IP, w server.ResponseSender, r *server.Request) error {
 
 // newIAAddr creates a IAAddr for a IANA using the specified IPv6 address,
 // and advertises it to a client.
-func newIAAddr(ia *dhcp6opts.IANA, ip net.IP, w server.ResponseSender, r *server.Request) error {
+func newIAAddr(ia *dhcp6opts.IANA, ip net.IP, w dhcp6server.ResponseSender, r *dhcp6server.Request) error {
 	// Send IPv6 address with 60 second preferred lifetime,
 	// 90 second valid lifetime, no extra options
 	iaaddr, err := dhcp6opts.NewIAAddr(ip, 60*time.Second, 90*time.Second, nil)
