+++ dhcp6server/server_test.go
@@ -1,4 +1,4 @@
-package server
+package dhcp6server
 
 import (
 	"bytes"
